mindmate-ai/code/memory-service/requirements.txt
# MindMate AI - Memory Service Dependencies
# Vector memory service with Pinecone and OpenAI embeddings

# Pinecone vector database
pinecone-client>=5.0.0

# OpenAI for embeddings
openai>=1.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Data handling
pydantic>=2.0.0
python-dateutil>=2.8.2

# Utilities
tenacity>=8.2.0
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
